requirements.txt
# Interfaz gráfica
customtkinter==5.2.1
pillow==10.1.0

# Procesamiento de imágenes
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3

# Manejo de Excel
openpyxl==3.1.2
pandas==2.1.3

# Utilidades
python-dateutil==2.8.2